+++ ingot-service/ingot-pms/ingot-pms-provider/src/main/java/com/ingot/cloud/pms/rest/v1/MenuApi.java
@@ -46,7 +46,7 @@ public class MenuApi extends BaseController {
     }
 
     @DeleteMapping("/{id}")
-    public R<?> removeById(@PathVariable Long id) {
+    public R<?> removeById(@PathVariable Integer id) {
         sysMenuService.removeMenuById(id);
         return ok();
     }
